+++ CHANGELOG.md
@@ -19,6 +19,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Prevent an overflowed grist gutter from negatively influencing ability to pick up grist
 - Fixed grist cache view in editmode using the limit of the wrong player
 - Cache limit will now be accounted for more often client-side
+- Clearer and more accurate grist toasts when at cache capacity
 
 ### Contributors for this release